chore: extract polling recorder
